[TEST] Jest setting #29
- Set env for test code
- Add simple test code
