+++ package.json
@@ -4,8 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon"
+    "dev": "nodemon",
+    "test": "jest --setupFiles dotenv/config --forceExit --detectOpenHandles"
   },
   "repository": {
     "type": "git",
@@ -22,6 +22,7 @@
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
+    "jest": "^27.3.1",
     "jsonwebtoken": "^8.5.1",
     "mysql2": "^2.3.0",
     "node-cron": "^3.0.0",
@@ -31,6 +32,7 @@
     "sequelize": "^6.6.5",
     "sequelize-cli": "^6.2.0",
     "socket.io": "^4.2.0",
+    "supertest": "^6.1.6",
     "swagger-cli": "^4.0.4",
     "swagger-ui-express": "^4.1.6",
     "uuid": "^8.3.2",
@@ -38,11 +40,13 @@
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
+    "@types/jest": "^27.0.2",
     "@types/jsonwebtoken": "^8.5.5",
     "@types/node-cron": "^3.0.0",
     "@types/passport": "^1.0.7",
     "@types/passport-jwt": "^3.0.6",
     "@types/passport-local": "^1.0.34",
+    "@types/supertest": "^2.0.11",
     "@types/swagger-ui-express": "^4.1.3",
     "@types/uuid": "^8.3.1",
     "@types/yamljs": "^0.2.31",
@@ -56,6 +60,7 @@
     "morgan": "^1.10.0",
     "nodemon": "^2.0.13",
     "prettier": "^2.4.1",
+    "ts-jest": "^27.0.7",
     "typescript": "^4.4.3"
   }
 }
