+++ src/app.ts
@@ -1,11 +1,7 @@
 import cors from 'cors';
 import express from 'express';
 import passport from 'passport';
-import path from 'path';
-import swaggerUi from 'swagger-ui-express';
-import YAML from 'yamljs';
 import passportConfig from './auth/passport';
-import sequelize from './models';
 import router from './routes';
 
 const app = express();
@@ -15,20 +11,6 @@ app.set('port', process.env.PORT || 3000);
 app.get('/', (req, res) => res.send('Hello Express'));
 
 app.use(cors());
-
-// swagger
-const swaggerYaml = YAML.load(path.join(__dirname, '../../swagger/swagger.yaml'));
-app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerYaml));
-
-// db connection
-sequelize
-  .sync({ force: false })
-  .then(() => {
-    console.log('db connect success!');
-  })
-  .catch((err: Error) => {
-    console.error(err);
-  });
 
 // body-parser
 app.use(express.json());
@@ -43,6 +25,4 @@ app.use('/api', router);
 // scheduler
 // initScheduler();
 
-app.listen(app.get('port'), () => {
-  console.log(app.get('port'), '번 포트에서 실행중');
-});
+export default app;
